+++ packages/expo-router/src/ContextNavigator.tsx
@@ -1,8 +1,9 @@
 import React, { useMemo } from "react";
 
 import { ContextNavigationContainer } from "./ContextNavigationContainer";
-import { RoutesContext, useRoutesContext } from "./context";
+import { RootRouteNodeContext, useRootRouteNodeContext } from "./context";
 import { getRoutes } from "./getRoutes";
+import { useTutorial } from "./onboard/useTutorial";
 import { RequireContext } from "./types";
 import { getQualifiedRouteComponent } from "./useScreens";
 
@@ -12,7 +13,7 @@ function useContextModuleAsRoutes(context: RequireContext) {
   return useMemo(() => getRoutes(context), [keys]);
 }
 
-function RoutesContextProvider({
+function RootRouteNodeProvider({
   context,
   children,
 }: {
@@ -21,46 +22,10 @@ function RoutesContextProvider({
 }) {
   const routes = useContextModuleAsRoutes(context);
   return (
-    <RoutesContext.Provider value={routes}>{children}</RoutesContext.Provider>
+    <RootRouteNodeContext.Provider value={routes}>
+      {children}
+    </RootRouteNodeContext.Provider>
   );
-}
-
-function isFunctionOrReactComponent(
-  Component: any
-): Component is React.ComponentType {
-  return (
-    !!Component &&
-    (typeof Component === "function" ||
-      Component?.prototype?.isReactComponent ||
-      Component.$$typeof === Symbol.for("react.forward_ref"))
-  );
-}
-
-/** Returns the Tutorial component if there are no React components exported as default from any files in the provided context module. */
-function useTutorial(context: RequireContext) {
-  if (process.env.NODE_ENV === "production") {
-    return null;
-  }
-
-  // eslint-disable-next-line react-hooks/rules-of-hooks
-  const keys = useMemo(() => context.keys(), [context]);
-  // eslint-disable-next-line react-hooks/rules-of-hooks
-  const hasAnyValidComponent = useMemo(() => {
-    for (const key of keys) {
-      // NOTE(EvanBacon): This should only ever occur in development as it breaks lazily loading.
-      const component = context(key)?.default;
-      if (isFunctionOrReactComponent(component)) {
-        return true;
-      }
-    }
-    return false;
-  }, [keys]);
-
-  if (hasAnyValidComponent) {
-    return null;
-  }
-
-  return require("./onboard/Tutorial").Tutorial;
 }
 
 export function ContextNavigator({ context }: { context: RequireContext }) {
@@ -70,16 +35,16 @@ export function ContextNavigator({ context }: { context: RequireContext }) {
   }
 
   return (
-    <RoutesContextProvider context={context}>
+    <RootRouteNodeProvider context={context}>
       <ContextNavigationContainer>
         <RootRoute />
       </ContextNavigationContainer>
-    </RoutesContextProvider>
+    </RootRouteNodeProvider>
   );
 }
 
 function RootRoute() {
-  const root = useRoutesContext();
+  const root = useRootRouteNodeContext();
 
   if (!root) {
     return null;
